+++ src/main/java/techcourse/myblog/controller/ArticleController.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.domain.Article;
-import techcourse.myblog.dto.ArticleDto;
+import techcourse.myblog.service.dto.ArticleDto;
 import techcourse.myblog.domain.ArticleRepository;
 
 @Controller
