+++ src/main/java/techcourse/myblog/domain/ArticleRepository.java
@@ -2,7 +2,7 @@ package techcourse.myblog.domain;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
-import techcourse.myblog.dto.ArticleDto;
+import techcourse.myblog.service.dto.ArticleDto;
 
 public interface ArticleRepository extends CrudRepository<Article, Long> {
     @Transactional
