+++ src/main/java/techcourse/myblog/support/config/WebConfig.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.support;
+package techcourse.myblog.support.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
