+++ src/main/java/techcourse/myblog/support/validation/UserGroups.java
@@ -1,4 +1,4 @@
-package techcourse.myblog;
+package techcourse.myblog.support.validation;
 
 public interface UserGroups {
     interface Edit{}
