+++ src/test/java/techcourse/myblog/service/dto/UserDtoTest.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.service.dto;
 
 
 import org.junit.jupiter.api.BeforeEach;
